Get initial configuration roughed out
